+++ igseq/igblast.py
@@ -70,6 +70,11 @@ def igblast(
     LOGGER.info("given colmap: %s", colmap)
     LOGGER.info("given extra args: %s", extra_args)
     LOGGER.info("given threads: %s", threads)
+    if species and not ref_paths:
+        # If only species is given, default to using all available reference
+        # sets for that species
+        ref_paths = [_fuzzy_species_match(species)]
+        LOGGER.info("inferred ref path: %s", ref_paths[0])
     attrs_list = vdj.parse_vdj_paths(ref_paths)
     for attrs in attrs_list:
         LOGGER.info("detected ref path: %s", attrs["path"])
@@ -104,12 +109,11 @@ def detect_organism(species_det, species=None):
         species = species_det.pop()
         LOGGER.info("detected species: %s", species)
     # match species names if needed
-    species_key = re.sub("[^a-z]", "", species.lower())
-    if species not in SPECIESMAP and species_key in SPECIESOTHER:
-        species_new = SPECIESOTHER[species_key]
+    species_new = _fuzzy_species_match(species)
+    if species_new != species:
         LOGGER.info(
-            "detected species as synonym: %s -> %s -> %s", species, species_key, species_new)
-        species = species_new
+            "detected species as synonym: %s -> %s", species, species_new)
+    species = species_new
     try:
         organism = SPECIESMAP[species]
     except KeyError as err:
@@ -117,6 +121,15 @@ def detect_organism(species_det, species=None):
         raise util.IgSeqError(f"species not recognized.  should be one of: {keys}") from err
     LOGGER.info("detected IgBLAST organism: %s", organism)
     return organism
+
+def _fuzzy_species_match(species):
+    """Fuzzy-match one of our species names"""
+    species_key = re.sub("[^a-z]", "", species.lower())
+    try:
+        return SPECIESOTHER[species_key]
+    except KeyError as err:
+        keys = str(SPECIESMAP.keys())
+        raise util.IgSeqError(f"species not recognized.  should be one of: {keys}") from err
 
 @contextmanager
 def run_igblast(
@@ -164,43 +177,51 @@ def run_igblast(
         # is handle proc.stdout.  When that block returns, we'll handle the
         # cleanup here.
         def stdin_writer():
-            # read whatever format from the query file, write FASTA to the
-            # igblastn proc's stdin
             try:
+                # read whatever format from the query file, write FASTA to the
+                # igblastn proc's stdin
                 with RecordReader(query_path, fmt_in, colmap) as reader, \
                     RecordWriter(proc.stdin, "fa", colmap) as writer:
                     for rec in reader:
                         writer.write(rec)
+            except BrokenPipeError:
+                pass
             # whatever else happens here, try to close the process' stdin, so
-            # we don't leave things hanging.  A ValueError is raised if the
-            # pipe is already closed, but if that's not the case, re-raise that
-            # exception if it occurs.
+            # we don't leave things hanging.
             finally:
                 try:
                     proc.stdin.close()
-                except ValueError:
-                    if not proc.stdin.closed:
-                        raise
+                except BrokenPipeError:
+                    pass
         def stderr_reader():
             # all this does is catch stderr line-by-line and write it to our
             # stderr.  this way it can be captured by contextlib.redirect_stderr.
-            # The warning in the contextlib docs about thatbeing "not suitable
+            # The warning in the contextlib docs about that being "not suitable
             # for use in library code and most threaded applications"
             # notwithstanding, it does seem to work here.
-            for line in proc.stderr:
-                sys.stderr.write(line)
+            try:
+                for line in proc.stderr:
+                    sys.stderr.write(line)
+            except BrokenPipeError:
+                pass
         # Start the threads for piping text to proc.stdin and from proc.stderr
         # (they'll block until there's something to do), yield the proc object,
         # and then handle cleanup once back here.
         pipe_threads = [
-            threading.Thread(target=stdin_writer),
-            threading.Thread(target=stderr_reader)]
+            threading.Thread(target=stdin_writer, name="stdin"),
+            threading.Thread(target=stderr_reader, name="stderr")]
         for thread in pipe_threads:
             thread.start()
-        yield proc
-        for thread in pipe_threads:
-            thread.join()
-        proc.wait()
+        try:
+            yield proc
+        # Whatever happens inside the yield, always clean up the process and
+        # its pipes.
+        finally:
+            if proc.returncode is None:
+                proc.terminate()
+            for thread in pipe_threads:
+                thread.join()
+            proc.wait()
         if proc.returncode:
             LOGGER.critical("%s exited with code %d", proc.args[0], proc.returncode)
             raise util.IgSeqError("IgBLAST crashed")